requirements.txt
Django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-smart-selects==1.7.2
djangorestframework-simplejwt==5.5.1

firebase-admin==7.2.0

celery==5.6.2
redis==7.3.0
eventlet==0.40.4

geopy==2.4.1

python-dotenv
requests

razorpay
